Fix a variety of bugs, move code around
